Fixed a couple of bugs relating to non accessible data.

+++ server/src/resolvers/Mutation/booking.js
@@ -1,5 +1,11 @@
+const { getUserId, AuthError } = require('../../utils')
 const booking = {
   async cancelBooking(parent, { input }, ctx) {
+    const userId = getUserId(ctx)
+    const bookee = await ctx.prisma.booking({ id: input.bookingId }).bookee()
+    if (bookee.id !== userId) {
+      throw new AuthError()
+    }
     const cancelledBooking = await ctx.prisma.deleteBooking({
       id: input.bookingId,
     })
